refactor(guidelines): only one level nesting
