+++ src/components/layer-manager/layer-item.element.ts
@@ -1,8 +1,8 @@
 import i18next from 'i18next'
 import { LitElement } from 'lit'
 import { property } from 'lit/decorators'
-import { i18nMixin } from '../../../mixins/i18n-lit-element'
-import { Layer } from '../../../states/map/map.state.model'
+import { i18nMixin } from '../../mixins/i18n-lit-element'
+import { Layer } from '../../states/map/map.state.model'
 
 export abstract class LayerItemElement<T = any> extends i18nMixin(LitElement) {
   @property()
