+++ src/components/layer-manager/layer-item/layer-item-background.element.ts
