+++ src/components/layer-manager/layer-item/layer-item.element.ts
