+++ src/components/layers-panel/layer-panel.element.ts
@@ -3,9 +3,9 @@ import { customElement, state } from 'lit/decorators'
 import i18next from 'i18next'
 import { i18nMixin } from '../../mixins/i18n-lit-element'
 
-import './catalog/remote-layers/remote-layers.element'
-import './catalog/catalog-tab.element'
-import './layer-manager/layer-manager.element'
+import '../remote-layers/remote-layers.element'
+import '../catalog/catalog-tab.element'
+import '../layer-manager/layer-manager.element'
 
 @customElement('lux-layer-panel')
 export class LayerPanelElement extends i18nMixin(LitElement) {
