+++ src/components/theme-selector/theme-selector.element.ts
@@ -1,8 +1,8 @@
 import { html, LitElement, TemplateResult } from 'lit'
 import { customElement, state } from 'lit/decorators'
 import { Subscription } from 'rxjs'
-import { ThemeNodeModel } from '../../../../services/themes/themes.model'
-import { themesService } from '../../../../services/themes/themes.service'
+import { ThemeNodeModel } from '../../services/themes/themes.model'
+import { themesService } from '../../services/themes/themes.service'
 import { themeSelectorService } from './theme-selector.service'
 
 import './theme-grid.element'
